a quick change to GLF to keep as much precision in our likelihoods as long as possible, before we put it into byte space.  Sanger was doing a diff at low coverage and noticed our calls didn't contain as much precision as theirs.  Updated the MD5 for unified genotyper output.
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@2644 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/utils/genotype/glf/GLFReader.java
@@ -108,7 +108,7 @@ public class GLFReader implements Iterator<GLFRecord> {
         short rmsMapping = inputBinaryCodec.readUByte();
         double[] lkValues = new double[LikelihoodObject.GENOTYPE.values().length];
         for (int x = 0; x < LikelihoodObject.GENOTYPE.values().length; x++) {
-            lkValues[x] = (inputBinaryCodec.readUByte() / GLFRecord.LIKELIHOOD_SCALE_FACTOR + min_lk);
+            lkValues[x] = ((double)inputBinaryCodec.readUByte() / GLFRecord.LIKELIHOOD_SCALE_FACTOR + (double)min_lk);
         }
         return new GLFSingleCall(referenceName, refBase, (int)(offset+currentLocation), readDepth, rmsMapping, lkValues);
     }
